Move /specification endpoint to separate controller and adjust unit tests accordingly.
