+++ src/main/resources/templates/index.html
@@ -37,7 +37,7 @@
             </button>
             <button
               class="flex w-64 transform items-center space-x-2 rounded border border-gray-400 bg-white px-4 py-2 font-semibold text-gray-800 shadow transition hover:bg-green-200"
-              hx-post="/loadSpecification?specification_type=example"
+              hx-post="/specification?specification_type=example"
               hx-target="#specificationDiv"
             >
               <i class="bi bi-file-earmark-text text-2xl text-green-600"></i>
@@ -92,7 +92,7 @@
       </div>
     </div>
     <form
-      hx-post="/loadSpecification?specification_type=pdf"
+      hx-post="/specification?specification_type=pdf"
       hx-target="#specificationDiv"
       hx-encoding="multipart/form-data"
       hx-trigger="change"
@@ -106,7 +106,7 @@
       />
     </form>
     <form
-      hx-post="/loadSpecification?specification_type=word"
+      hx-post="/specification?specification_type=word"
       hx-target="#specificationDiv"
       hx-encoding="multipart/form-data"
       hx-trigger="change"
