+++ src/main/resources/templates/navbar-fragment.html
@@ -5,7 +5,7 @@
         <!-- Logo -->
         <div class="flex items-center">
           <img th:src="@{/img/spontecular-logo.png}" class="h-12" alt="" />
-          <a href="#" class="ml-2 text-2xl font-bold text-gray-800"
+          <a href="/" class="ml-2 text-2xl font-bold text-gray-800"
             >SpOntecular</a
           >
         </div>
