+++ src/test/java/com/example/spontecular/controller/AppControllerTest.java
@@ -34,9 +34,6 @@ class AppControllerTest {
     @MockBean
     private JenaService jenaService;
 
-    @MockBean
-    private SpecificationService specificationService;
-
     @Test
     void shouldLoadIndex() throws Exception {
         MvcResult mvcResult = mockMvc.perform(get("/"))
@@ -44,50 +41,6 @@ class AppControllerTest {
                 .andReturn();
 
         assertThat(mvcResult.getModelAndView().getViewName()).isEqualTo("index");
-    }
-
-    @Test
-    void shouldLoadExampleSpecification() throws Exception {
-        when(specificationService.loadExampleSpecification()).thenReturn("example specification");
-
-        MvcResult mvcResult = mockMvc.perform(post("/loadSpecification")
-                        .param("specification_type", "example"))
-                .andExpect(status().isOk())
-                .andReturn();
-
-        assertThat(mvcResult.getModelAndView().getViewName())
-                .isEqualTo("fragments :: specificationFragment");
-
-        assertThat(mvcResult.getModelAndView().getModel().get("specification"))
-                .isEqualTo("example specification");
-    }
-
-    @Test
-    void shouldLoadPdfSpecification() throws Exception {
-        MockMultipartFile pdfFile = new MockMultipartFile("pdfFile", "test.pdf", "application/pdf", "PDF content".getBytes());
-        when(specificationService.loadPdfSpecification(any())).thenReturn("PDF Spec");
-        MvcResult result = mockMvc.perform(multipart("/loadSpecification")
-                        .file(pdfFile)
-                        .param("specification_type", "pdf"))
-                .andExpect(status().isOk())
-                .andReturn();
-
-        assertThat(result.getModelAndView().getViewName()).isEqualTo("fragments :: specificationFragment");
-        assertThat(result.getModelAndView().getModel().get("specification")).isEqualTo("PDF Spec");
-    }
-
-    @Test
-    void shouldLoadWordSpecification() throws Exception {
-        MockMultipartFile wordFile = new MockMultipartFile("wordFile", "test.docx", "application/vnd.openxmlformats-officedocument.wordprocessingml.document", "Word content".getBytes());
-        when(specificationService.loadWordSpecification(any())).thenReturn("Word Spec");
-        MvcResult result = mockMvc.perform(multipart("/loadSpecification")
-                        .file(wordFile)
-                        .param("specification_type", "word"))
-                .andExpect(status().isOk())
-                .andReturn();
-
-        assertThat(result.getModelAndView().getViewName()).isEqualTo("fragments :: specificationFragment");
-        assertThat(result.getModelAndView().getModel().get("specification")).isEqualTo("Word Spec");
     }
 
     @Test
